% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{Addranks}
\alias{Addranks}
\title{Add classification levels}
\usage{
Addranks(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{Allindices}}.}
}
\value{
A datafrom with columns containing classification information, i.e. family, order, class, phylum, kingdom.
}
\description{
This function adds classiication levels to the datafrom created from the Index Summary function.
}
\examples{
data(Woylie)
Allindices(Woylie, genus = "genus_name", species = "species_name")
Addranks(Woylie)

}
