% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{ARRatio}
\alias{ARRatio}
\title{Article:Review ratio}
\usage{
ARRatio(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
A character value of the percentage ratio of the number of articles and reviews.
}
\description{
This function calculates the percentage ratio of article:rerview.
}
\examples{
data(Woylie)
ARRatio(Woylie)

}
