% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{Count}
\alias{Count}
\title{Search count of literature}
\usage{
Count(db, search, level, genus, species = NULL, synonyms, additionalkeywords)
}
\arguments{
\item{db}{Literature database. Scopus ("scopus"), Web of Science ("wos"), or Base ("base").}

\item{search}{Search fields. Title only ("t") or title, abstract, and keywords ("tak").}

\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{synonyms}{Alternate species names.}

\item{additionalkeywords}{Optional search terms.}
}
\value{
Search count of the genus or species with the given \code{genus} and/or \code{species}.
}
\description{
This function counts the total number of search results from Scopus, Web of Science, or BASE.
A check will be conducted via \code{\link[taxize]{gnr_resolve}} to validate the genus and species names.
}
\examples{
\dontrun{
Count(db = "scopus",
      search = "t",
      genus = "Osphranter", species = "rufus")
}
\dontrun{
Count(db = "scopus",
      search = "t",
      genus = "Osphranter", species = "rufus",
      synonyms = "Macropus rufus",
      additionalkeywords = "conserv*")
}
}
