% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{CountDomain}
\alias{CountDomain}
\title{Restricted domain}
\usage{
CountDomain(genus, species, keyword, APIkey, datatype = "application/xml")
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{keyword}{Restricted domain.}

\item{APIkey}{Scopus API key needed to access and download data from their database.}

\item{datatype}{Formats the URL to be sent to the API. The default is "application/xml".}
}
\value{
Search count of a specific restricted domain of a species with the given \code{genus} and \code{species}.
}
\description{
Add a keyword in the search string to restrict the domain of a search.
It counts the publications with the binomial name and keyword in the title, abstract and keywords.
}
\examples{
\dontrun{
CountDomain("Bettongia", "penicillata", "conserv*", "myAPI")

#lower case letter in genus is also accepted and will return identical results

CountDomain("bettongia", "penicillata", "conserv*", "myAPI")
}
}
