% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{CountGenusTAK}
\alias{CountGenusTAK}
\title{CountSpTAK wrapper}
\usage{
CountGenusTAK(db, genus, synonyms, additionalkeywords)
}
\arguments{
\item{db}{Literature database. Scopus ("scopus"), Web of Science ("wos"), or Base ("base").}

\item{genus}{Genus classification from the binomial name.}

\item{synonyms}{Alternate genus names.}

\item{additionalkeywords}{Optional search terms.}
}
\value{
Search count of the genus with the given \code{genus}.
}
\description{
This is a wrapper function for \code{\link{CountGenusTAK_scopus}}, \code{\link{CountGenusTAK_wos}}, and \code{\link{CountGenusTAK_base}}.
}
\examples{
\dontrun{
CountGenusTAK("scopus", genus = "Osphranter")
}
\dontrun{
CountGenusTAK("scopus", genus = "Osphranter", synonyms = "Macropus", additionalkeywords = "conserv*")
}
}
