% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{CountGenusTAK_lens}
\alias{CountGenusTAK_lens}
\title{Search count from Lens - title, abstract and author keywords}
\usage{
CountGenusTAK_lens(genus, synonyms, additionalkeywords, size = 50000)
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{synonyms}{Alternate genus names.}

\item{additionalkeywords}{Optional search terms.}

\item{size}{Maximum number of documents that can be downloaded depending on the users token. Default is set to 50,000 for subscribers, the alternative is 1,000 for non-subscribers.}
}
\value{
Search count of the genus with the given \code{genus}.
}
\description{
This function counts the total number of search results.
It counts the publications with the genus name in the title, abstract and author keywords.
A check will be conducted via \code{\link[taxize]{gnr_resolve}} to validate the genus name.
}
\examples{
\dontrun{
CountGenusTAK_lens("Bettongia")

#lower case letter in genus is also accepted and will return identical results

CountGenusTAK_lens("bettongia")
}
\dontrun{
CountGenusTAK_lens("Bettongia", "conserv*")

#lower case letter in genus is also accepted and will return identical results

CountGenusTAK_lens("bettongia", "conserv*")
}
}
\references{
Chamberlain, S. & Szocs, E. (2013). taxize - taxonomic search and retrieval in R. \emph{F1000Research, 2}, 191.
}
