% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{CountGenusTAK_wos}
\alias{CountGenusTAK_wos}
\title{Search count from Web of Science - title, abstract and author keywords}
\usage{
CountGenusTAK_wos(genus, synonyms, additionalkeywords)
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{synonyms}{Alternate genus names.}

\item{additionalkeywords}{Optional search terms.}
}
\value{
Search count of the genus with the given \code{genus}.
}
\description{
This function counts the total number of search results.
It counts the publications with the genus name in the title, abstract and author keywords.
A check will be conducted via \code{\link[taxize]{gnr_resolve}} to validate the genus name.
}
\examples{
\dontrun{
CountGenusTAK_wos("Bettongia")

#lower case letter in genus is also accepted and will return identical results

CountGenusTAK_wos("bettongia")
}
\dontrun{
CountGenusTAK_wos("Bettongia", "conserv*")

#lower case letter in genus is also accepted and will return identical results

CountGenusTAK_wos("bettongia", "conserv*")
}
}
\references{
Chamberlain, S. & Szocs, E. (2013). taxize - taxonomic search and retrieval in R. \emph{F1000Research, 2}, 191.
}
