% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{CountSpT_base}
\alias{CountSpT_base}
\title{Title}
\usage{
CountSpT_base(genus, species, synonyms, additionalkeywords)
}
\arguments{
\item{genus}{}

\item{species}{}

\item{synonyms}{}

\item{additionalkeywords}{}
}
\value{

}
\description{
Title
}
