% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{CountSpT_lens}
\alias{CountSpT_lens}
\title{Search count from Lens - title only}
\usage{
CountSpT_lens(genus, species, synonyms, additionalkeywords, token)
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{synonyms}{Alternate species names.}

\item{additionalkeywords}{Optional search terms.}

\item{token}{Lens token needed to access and download data from their database.}
}
\value{
Search count of the species with the given \code{genus} and \code{species}.
}
\description{
This function counts the total number of search results.
It counts the publications with the binomial name in the title only.
A check will be conducted via \code{\link[taxize]{gnr_resolve}} to validate the genus and species names.
}
\examples{
\dontrun{
CountSpT_lens("Bettongia", "penicillata")

#lower case letter in genus is also accepted and will return identical results

CountSpT_lens("bettongia", "penicillata")
}
\dontrun{
CountSpT_lens("Bettongia", "penicillata", "conserv*")

#lower case letter in genus is also accepted and will return identical results

CountSpT_lens("bettongia", "penicillata", "conserv*")
}
}
\references{
Chamberlain, S. & Szocs, E. (2013). taxize - taxonomic search and retrieval in R. \emph{F1000Research, 2}, 191.
}
