% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{CountSpTAK}
\alias{CountSpTAK}
\title{CountSpTAK wrapper}
\usage{
CountSpTAK(db, genus, species, synonyms, additionalkeywords, size = 50000)
}
\arguments{
\item{db}{Literature database. Scopus ("scopus") or Web of Science ("wos") or Base ("base") or Lens ("lens").}

\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{synonyms}{Alternate species names.}

\item{additionalkeywords}{Optional search terms.}

\item{size}{Maximum number of documents that can be downloaded depending on the users token. Default is set to 50,000 for subscribers, the alternative is 1,000 for non-subscribers. Lens only.}
}
\value{
Search count of the species with the given \code{genus} and \code{species}.
}
\description{
This is a wrapper function for \code{\link{CountSpTAK_scopus}}, \code{\link{CountSpTAK_wos}}, \code{\link{CountSpTAK_base}}, and \code{\link{CountSpTAk_lens}}.
}
\examples{
\dontrun{
CountSpTAK("scopus", genus = "Osphranter", species = "rufus")
}
\dontrun{
CountSpTAK("scopus", genus = "Osphranter", species = "rufus", synonyms = "Macropus rufus", additionalkeywords = "conserv*")
}
}
