% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{CountSpTAK_lens}
\alias{CountSpTAK_lens}
\title{Search count from Lens - title, abstract and author keywords}
\usage{
CountSpTAK_lens(
  genus,
  species,
  synonyms,
  additionalkeywords,
  token,
  size = 50000
)
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{synonyms}{Alternate species names.}

\item{additionalkeywords}{Optional search terms.}

\item{token}{Lens token needed to access and download data from their database.}

\item{size}{Maximum number of documents that can be downloaded depending on the users token. Default is set to 50,000 for subscribers, the alternative is 1,000 for non-subscribers.}
}
\value{
Search count of the species with the given \code{genus} and \code{species}.
}
\description{
This function counts the total number of search results.
It counts the publications with the binomial name in the title, abstract and author keywords.
A check will be conducted via \code{\link[taxize]{gnr_resolve}} to validate the genus and species names.
}
\examples{
\dontrun{
CountSpTAK_lens("Bettongia", "penicillata")

#lower case letter in genus is also accepted and will return identical results

CountSpTAK_lens("bettongia", "penicillata")
}
\dontrun{
CountSpTAK_lens("Bettongia", "penicillata", "conserv*")

#lower case letter in genus is also accepted and will return identical results

CountSpTAK_lens("bettongia", "penicillata", "conserv*")
}
}
\references{
Chamberlain, S. & Szocs, E. (2013). taxize - taxonomic search and retrieval in R. \emph{F1000Research, 2}, 191.
}
