% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{CountSpTAK_scopus}
\alias{CountSpTAK_scopus}
\title{Search count from Scopus - title, abstract, and keywords}
\usage{
CountSpTAK_scopus(
  genus,
  species,
  synonyms,
  additionalkeywords,
  datatype = "application/xml"
)
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{synonyms}{Alternate species names.}

\item{additionalkeywords}{Optional search terms.}

\item{datatype}{Formats the URL to be sent to the API. The default is "application/xml".}
}
\value{
Search count of the species with the given \code{genus} and \code{species}.
}
\description{
This function counts the total number of search results.
It counts the publications with the binomial name in the title, abstract, and keywords.
A check will be conducted via \code{\link[taxize]{gnr_resolve}} to validate the genus and species names.
}
\examples{
\dontrun{
CountSpTAK_scopus("Bettongia", "penicillata")

#lower case letter in genus is also accepted and will return identical results

CountSpTAK_scopus("bettongia", "penicillata")
}
\dontrun{
CountSpTAK_scopus("Bettongia", "penicillata")

#lower case letter in genus is also accepted and will return identical results

CountSpTAK_scopus("bettongia", "penicillata")
}
}
\references{
Chamberlain, S. & Szocs, E. (2013). taxize - taxonomic search and retrieval in R. \emph{F1000Research, 2}, 191.
}
