% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{CountSpTexp}
\alias{CountSpTexp}
\title{Expanded string}
\usage{
CountSpTexp(
  genus,
  species,
  additionalkeywords = " ",
  APIkey,
  datatype = "application/xml"
)
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{additionalkeywords}{Additional search terms.}

\item{APIkey}{Scopus API key needed to access and download data from their database.}

\item{datatype}{Formats the URL to be sent to the API. The default is "application/xml".}
}
\value{
Search count of a specific search string of a species with the given \code{genus} and \code{species}.
}
\description{
Expansion of \code{\link{CountSpT}}, adds other search terms the search string.
It counts the publications with the binomial name and keyword in the title only.
}
\examples{
\dontrun{
CountSpTexp("Bettongia", "penicillata", "conserv*", "myAPI")

#lower case letter in genus is also accepted and will return identical results

CountSpTexp("bettongia", "penicillata", "conserv*", "myAPI")
}
}
