% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{extractcontent}
\alias{extractcontent}
\title{Extract content}
\usage{
extractcontent(search.string, datatype = "application/xml")
}
\arguments{
\item{search.string}{Search string with Boolean operators or Scopus advanced search.}

\item{datatype}{Formats the URL to be sent to the API. The default is "application/xml".}
}
\value{
A list of entries of the search from Scopus.
}
\description{
Extract citation data from Scopus.
}
\examples{
extractcontent("TITLE-ABS-KEY("bettongia penicillata") AND DOCTYPE(ar OR re)")

}
