% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{extractXML}
\alias{extractXML}
\title{Extract XML list into a dataframe.}
\usage{
extractXML(theFile)
}
\arguments{
\item{theFile}{The file to be converted.}
}
\value{
A converted dataframe generated from \code{\link{extractcontent}}.
}
\description{
Extract XML list into a dataframe.
}
\examples{
extractXML(SpeciesXML)

}
