% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{Fetch}
\alias{Fetch}
\title{Fetch citation records}
\usage{
Fetch(
  db,
  search,
  genus,
  species = NULL,
  synonyms,
  additionalkeywords,
  language = 0
)
}
\arguments{
\item{db}{Literature database. Scopus ("scopus"), Web of Science ("wos"), or Base ("base").}

\item{search}{Search fields. Title only ("t") or title, abstract, and keywords ("tak").}

\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{synonyms}{Alternate species names.}

\item{additionalkeywords}{Optional search terms.}

\item{language}{Language of the paper; default is 0, enter 1 to retrieve the variable. Scopus only.}
}
\value{
A dataframe of the genus' or species' citation records with the given \code{genus} and/or \code{species}.
}
\description{
This function fetches citation information from Scopus, Web of Science, or BASE.
Duplicates are to be removed by the user after fetching the data.
}
\examples{
\dontrun{
Fetch(db = "scopus",
      search = "t",
      genus = "Osphranter", species = "rufus")
}
\dontrun{
Fetch(db = "scopus",
      search = "t",
      genus = "Osphranter", species = "rufus",
      synonyms = "Macropus rufus",
      additionalkeywords = "conserv*")
}
}
