% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{FetchGenusT_scopus}
\alias{FetchGenusT_scopus}
\title{Fetch data from Scopus - title only}
\usage{
FetchGenusT_scopus(genus, synonyms, additionalkeywords, language = 0)
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{synonyms}{Alternate genus names.}

\item{additionalkeywords}{Optional search terms.}

\item{language}{Language of the paper; default is 0, enter 1 to retrieve the variable.}
}
\value{
A dataframe of the genus' citation records with the given \code{genus}.
}
\description{
This function fetches citation information from Scopus using genus name found in the title of the publications.
Duplicates are to be removed by the user after fetching the data.
}
\examples{
\dontrun{
FetchGenusT_scopus("Bettongia")

#lower case letter in genus is also accepted and will return identical results

FetchGenusT_scopus("bettongia")
}
\dontrun{
FetchGenusT_scopus("Bettongia", "conserv*")

#lower case letter in genus is also accepted and will return identical results

FetchGenusT_scopus("bettongia", "conserv*")
}
}
