% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{FetchGenusTAK}
\alias{FetchGenusTAK}
\title{FetchGenusTAK wrapper}
\usage{
FetchGenusTAK(db, genus, synonyms, additionalkeywords, language = 0)
}
\arguments{
\item{db}{Literature database. Scopus ("scopus") or Web of Science ("wos").}

\item{genus}{Genus classification from the binomial name.}

\item{synonyms}{Alternate genus names.}

\item{additionalkeywords}{Optional search terms.}

\item{language}{Language of the paper; default is 0, enter 1 to retrieve the variable. Scopus only.}
}
\value{
A dataframe of the genus' citation records with the given \code{genus}.
}
\description{
This is a wrapper function for \code{\link{FetchGenusTAK_scopus}}, and \code{\link{FetchGenusTAK_wos}}.
}
\examples{
\dontrun{
FetchGenusTAK("scopus", genus = "Osphranter")
}
\dontrun{
FetchGenusTAK("scopus", genus = "Osphranter", synonyms = "Macropus", additionalkeywords = "conserv*")
}
}
