% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{FetchSpT_scopus}
\alias{FetchSpT_scopus}
\title{Fetch data from Scopus - title only}
\usage{
FetchSpT_scopus(genus, species, synonyms, additionalkeywords, language = 0)
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{synonyms}{Alternate species names.}

\item{additionalkeywords}{Optional search terms.}

\item{language}{Language of the paper; default is 0, enter 1 to retrieve the variable.}
}
\value{
A dataframe of the species' citation records with the given \code{genus} and \code{species}.
}
\description{
This function fetches citation information from Scopus using genus and species name found in the title of the publications.
Duplicates are to be removed by the user after fetching the data.
}
\examples{
\dontrun{
FetchSpT_scopus("Bettongia", "penicillata")

#lower case letter in genus is also accepted and will return identical results

FetchSpT_scopus("bettongia", "penicillata")
}
\dontrun{
FetchSpT_scopus("Bettongia", "penicillata", "conserv*")

#lower case letter in genus is also accepted and will return identical results

FetchSpT_scopus("bettongia", "penicillata", "conserv*")
}
}
