% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{FetchSpT_wos}
\alias{FetchSpT_wos}
\title{Fetch data from Web of Science - title only}
\usage{
FetchSpT_wos(genus, species, synonyms, additionalkeywords)
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{synonyms}{Alternate species names.}

\item{additionalkeywords}{Optional search terms.}
}
\value{
A dataframe of the species' citation records with the given \code{genus} and \code{species}.
}
\description{
This function fetches citation information from Web of Science using genus and species name found in the title of the publications.
Duplicates are to be removed by the user after fetching the data.
}
\examples{
\dontrun{
FetchSpT_wos("Bettongia", "penicillata")

#lower case letter in genus is also accepted and will return identical results

FetchSpT_wos("bettongia", "penicillata")
}
\dontrun{
FetchSpT_wos("Bettongia", "penicillata", "conserv*")

#lower case letter in genus is also accepted and will return identical results

FetchSpT_wos("bettongia", "penicillata", "conserv*")
}
}
