% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{FetchSpTAK}
\alias{FetchSpTAK}
\title{FetchSpTAK wrapper}
\usage{
FetchSpTAK(
  db,
  genus,
  species,
  synonyms,
  additionalkeywords,
  language = 0,
  size = 50000
)
}
\arguments{
\item{db}{Literature database. Scopus ("scopus") or Web of Science ("wos") or Lens ("lens").}

\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{synonyms}{Alternate species names.}

\item{additionalkeywords}{Optional search terms.}

\item{language}{Language of the paper; default is 0, enter 1 to retrieve the variable. Scopus only.}

\item{size}{Maximum number of documents that can be downloaded depending on the users token. Default is set to 50,000 for subscribers, the alternative is 1,000 for non-subscribers. Lens only.}
}
\value{
A dataframe of the species' citation records with the given \code{genus} and \code{species}.
}
\description{
This is a wrapper function for \code{\link{FetchSpTAK_scopus}}, \code{\link{FetchSpTAK_wos}}, \code{\link{FetchSpTAK_base}}, and \code{\link{FetchSpTAK_lens}}.
}
\examples{
\dontrun{
FetchSpTAK("scopus", genus = "Osphranter", species = "rufus")
}
\dontrun{
FetchSpTAK("scopus", genus = "Osphranter", species = "rufus", synonyms = "Macropus rufus", additionalkeywords = "conserv*")
}
}
