% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{FetchSpTAK_lens}
\alias{FetchSpTAK_lens}
\title{Fetch data from Lens - title, abstract, and keywords.}
\usage{
FetchSpTAK_lens(genus, species, synonyms, additionalkeywords, token)
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{synonyms}{Alternate species names.}

\item{additionalkeywords}{Optional search terms.}

\item{token}{Lens token needed to access and download data from their database.}
}
\value{
A dataframe of the species' citation records with the given \code{genus} and \code{species}.
}
\description{
This function fetches citation information from Lens using genus and species name found in the title, abstract, and keywords of the publications.
Duplicates are to be removed by the user after fetching the data.
}
\examples{
\dontrun{
FetchSpTAK_lens("Bettongia", "penicillata", token = "mytoken")

#lower case letter in genus is also accepted and will return identical results

FetchSpTAK_lens("bettongia", "penicillata", token = "mytoken")
}
\dontrun{
FetchSpTAK_lens("Bettongia", "penicillata", "conserv*", token = "mytoken")

#lower case letter in genus is also accepted and will return identical results

FetchSpTAK_lens("bettongia", "penicillata", "conserv*", token = "mytoken")
}
}
