% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{getYear}
\alias{getYear}
\title{Extract year}
\usage{
getYear(data, genus, species)
}
\arguments{
\item{data}{Output from any of the fetch function.}

\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}
}
\value{
A dataframe with the year and frequency of the publications
}
\description{
Extracts the year of each publication of the output from any of the Fetch functions and counts the number of publications each year.
}
\examples{
\dontrun{
Woylie <- Fetch(db = "scopus",
                search = "tak",
                level = "species",
                genus = "Bettongia", species = "penicillata")
}
getYear(data = Woylie,
        genus = "Bettongia", species = "penicillata")

}
