% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{HAfterdate}
\alias{HAfterdate}
\title{This function calculates the h-index using a given date up till the newest record.}
\usage{
HAfterdate(data, date)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}

\item{date}{The lower limit of the timeframe.}
}
\value{
H-index of the given time period.
}
\description{
This function calculates the h-index using a given date up till the newest record.
}
\examples{
HAfterdate(SpeciesData, "2000-01-01")

}
