% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\docType{data}
\name{languages}
\alias{languages}
\title{languages}
\format{
A data frame :
\describe{
  \item{language}{names of languages}
  ...
}
}
\source{
somewhere?
}
\usage{
languages
}
\description{
A dataset with some literature on languages
}
\keyword{datasets}
