% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{plotPub}
\alias{plotPub}
\title{Publication plot}
\usage{
plotPub(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{getYear}}.}
}
\value{
ggplot
}
\description{
Plots the publication by year of a single species or combined.
}
\examples{
\dontrun{
Woylie <- Fetch(db = "scopus",
                search = "tak",
                level = "species",
                genus = "Bettongia", species = "penicillata")
Quokka <- Fetch(db = "scopus",
                search = "tak",
                level = "species",
                genus = "Setonix", species = "brachyurus")
Platypus <- Fetch(db = "scopus",
                  search = "tak",
                  level = "species",
                  genus = "Ornithorhynchus", species = "anatinus")
Koala <- Fetch(db = "scopus",
               search = "tak",
               level = "species",
               genus = "Phascolarctos", species = "cinereus")
}
extract_year_W <- getYear(data = Woylie,
                          genus = "Bettongia", species = "penicillata")
extract_year_Q <- getYear(data = Quokka,
                          genus = "Setonix", species = "brachyurus")
extract_year_P <- getYear(data = Platypus,
                          genus = "Ornithorhynchus", species = "anatinus")
extract_year_K <- getYear(data = Koala,
                          genus = "Phascolarctos", species = "cinereus")
Combine_pub <- rbind(extract_year_W, extract_year_Q, extract_year_P, extract_year_K)
plotPub(Combine_pub)

}
