% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{searchCount}
\alias{searchCount}
\title{This function counts the total number of search results.}
\usage{
searchCount(genus, species, APIkey, datatype = "application/xml")
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{APIkey}{Scopus API key needed to access and download data from their database.}

\item{datatype}{Formats the URL to be sent to the API. The default is "application/xml".}
}
\value{
Search count of the species with the given \code{genus} and \code{species}.
}
\description{
This function counts the total number of search results.
}
\examples{
searchCount("Bettongia", "penicillata", "442b9048417ef20cf680a0ae26ee4d86")

#lower case letter in genus is also accepted and will return identical results

searchCount("bettongia", "penicillata", "442b9048417ef20cf680a0ae26ee4d86")

}
