% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{SourceType}
\alias{SourceType}
\title{Source type}
\usage{
SourceType(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{Fetch}}.}
}
\value{
A dataframe with each document and their counts.
}
\description{
This function calculates the total number of items for each document type.
}
\examples{
\dontrun{
Woylie <- Fetch(db = "scopus",
                search = "tak",
                genus = "Bettongia",
                species = "penicillata")
}
SourceType(Woylie)

}
