% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{SpH5}
\alias{SpH5}
\title{Species h5 index}
\usage{
SpH5(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{Fetch}}.}
}
\value{
H5 index.
}
\description{
This function calculates the h-index of a species in the past 5 years.
}
\examples{
\dontrun{
Woylie <- Fetch(db = "scopus",
                search = "tak",
                genus = "Bettongia", 
                species = "penicillata")
}
SpH5(Woylie)

}
\references{
Suzuki, H. (2012). \emph{Google Scholar Metrics for Publications.} Retrieved from \url{https://scholar.googleblog.com/2012/04/google-scholar-metrics-for-publications.html}.
}
