% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{SpHGrowth}
\alias{SpHGrowth}
\title{H-index growth}
\usage{
SpHGrowth(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
A dataframe of the cumulative h-index.
}
\description{
This function calculates the cumulative h-index overtime by year.
}
\examples{
data(Woylie)
SpHGrowth(Woylie)

}
