% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{SpHindex}
\alias{SpHindex}
\title{Species h-index}
\usage{
SpHindex(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
H-index.
}
\description{
This function calculates the h-index of a species.
}
\examples{
SpHindex(SpeciesData)

}
