% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{SpHYear}
\alias{SpHYear}
\title{H-index by year}
\usage{
SpHYear(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\description{
This function calculates the h-index by year.
}
\examples{
SpHGrowth(data)

}
