% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{Spi10}
\alias{Spi10}
\title{This function calculates the i10 index of a species.
i10 index counts all of the publications with 10 or more citations.}
\usage{
Spi10(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
i10 index.
}
\description{
This function calculates the i10 index of a species.
i10 index counts all of the publications with 10 or more citations.
}
\examples{
Spi10(SpeciesData)

}
