% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{SpMindex}
\alias{SpMindex}
\title{Species m-index}
\usage{
SpMindex(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
M-index.
}
\description{
This function calculates the m-index of  species.
M-index uses the h-index and divides it by the number of years of activity.
}
\examples{
data(Woylie)
SpMindex(Woylie)

}
\references{
Hirsch, J. (2005). An index to quantify an individual's scientific research output. Proceedings of the National Academy of Sciences of the United States of America, 102(46), 16569-16572.
}
