% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{TotalArt}
\alias{TotalArt}
\title{Total Article}
\usage{
TotalArt(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
An integer of the total number of articles.
}
\description{
This function calculates the total number of articles.
}
\examples{
TotalArt(SpeciesData)

}
