% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{TotalJournals}
\alias{TotalJournals}
\title{This function calculates the total number of journals.}
\usage{
TotalJournals(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
An integer of the total number of journals.
}
\description{
This function calculates the total number of journals.
}
\examples{
TotalJournals(SpeciesData)

}
