% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{TotalRev}
\alias{TotalRev}
\title{Total reviews}
\usage{
TotalRev(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
An integer of the total number of reviews.
}
\description{
This function calculates the total number of reviews.
}
\examples{
data(Woylie)
TotalRev(Woylie)

}
