% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{YearsPublishing}
\alias{YearsPublishing}
\title{Years since first publication}
\usage{
YearsPublishing(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{Fetch}}.}
}
\value{
Number of years.
}
\description{
The number of years since the first publication in relation to the species.
}
\examples{
\dontrun{
Woylie <- Fetch(db = "scopus",
                search = "tak",
                genus = "Bettongia",
                species = "penicillata")
}
YearsPublishing(Woylie)

}
