% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{Allindices}
\alias{Allindices}
\title{Index summary}
\usage{
Allindices(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
A datarame of all of the indices in the package.
}
\description{
This function returns a dataframe of the summary of all of the indices.
}
\examples{
Allindices(data)

}
