% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{CountSpT}
\alias{CountSpT}
\title{CountSpT wrapper}
\usage{
CountSpT(db, genus, species, synonyms, additionalkeywords)
}
\arguments{
\item{db}{Literature database. Scopus ("scopus"), Web of Science ("wos"), or Base ("base").}

\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{synonyms}{Alternate species names.}

\item{additionalkeywords}{Optional search terms.}
}
\value{
Search count of the species with the given \code{genus} and \code{species}.
}
\description{
This is a wrapper function for \code{\link{CountSpT_scopus}}, \code{\link{CountSpT_wos}}, and \code{\link{CountSpT_base}}.
}
\examples{
\dontrun{
CountSpT("scopus", genus = "Osphranter", species = "rufus")
}
\dontrun{
CountSpT("scopus", genus = "Osphranter", species = "rufus", synonyms = "Macropus rufus", additionalkeywords = "conserv*")
}
}
