% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{CountSpT}
\alias{CountSpT}
\title{Search count - title only}
\usage{
CountSpT(genus, species, APIkey, datatype = "application/xml")
}
\arguments{
\item{genus}{}

\item{species}{}

\item{APIkey}{}

\item{datatype}{}
}
\value{

}
\description{
This function counts the total number of search results.
It counts only the publications with the binomial name in their title.
}
