% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{CountSpT}
\alias{CountSpT}
\title{Search count - title only}
\usage{
CountSpT(genus, species, APIkey, datatype = "application/xml")
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{APIkey}{Scopus API key needed to access and download data from their database.}

\item{datatype}{Formats the URL to be sent to the API. The default is "application/xml".}
}
\value{
Search count of the species with the given \code{genus} and \code{species}.
}
\description{
This function counts the total number of search results.
It counts only the publications with the binomial name in their title.
A check will be conducted via \code{\link[taxize]{gnr_resolve}} to validify the genus and species names.
}
\examples{
\dontrun{
CountSpT("Bettongia", "penicillata", "myAPI")

#lower case letter in genus is also accepted and will return identical results

CountSpT("Bettongia", "penicillata", "myAPI")
}
}
\references{
Chamberlain, S. & Szocs, E. (2013). taxize - taxonomic search and retrieval in R. \emph{F1000Research, 2}, 191.
}
