% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{CountSpTAK}
\alias{CountSpTAK}
\title{CountSpTAK wrapper}
\usage{
CountSpTAK(db, genus, species, synonyms, additionalkeywords)
}
\arguments{
\item{db}{Literature database. Scopus ("scopus"), Web of Science ("wos"), or Base ("base").}

\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{synonyms}{Alternate species names.}

\item{additionalkeywords}{Optional search terms.}
}
\value{
Search count of the species with the given \code{genus} and \code{species}.
}
\description{
This is a wrapper function for \code{\link{CountSpTAK_scopus}}, \code{\link{CountSpTAK_wos}}, and \code{\link{CountSpTAK_base}}.
}
\examples{
\dontrun{
CountSpTAK("scopus", genus = "Osphranter", species = "rufus")
}
\dontrun{
CountSpTAK("scopus", genus = "Osphranter", species = "rufus", synonyms = "Macropus rufus", additionalkeywords = "conserv*")
}
}
