% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{CountSpTAK}
\alias{CountSpTAK}
\title{Search count - title, abstract and keywords}
\usage{
CountSpTAK(genus, species, APIkey, datatype = "application/xml")
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{APIkey}{Scopus API key needed to access and download data from their database.}

\item{datatype}{Formats the URL to be sent to the API. The default is "application/xml".}
}
\value{
Search count of the species with the given \code{genus} and \code{species}.
}
\description{
This function counts the total number of search results.
It counts the publications with the binomial name in the title, abstract and keywords.
A check will be conducted via \code{taxize} to validify the genus and species names.
}
\examples{
CountSpTAK("Bettongia", "penicillata", "442b9048417ef20cf680a0ae26ee4d86")

#lower case letter in genus is also accepted and will return identical results

CountSpTAK("bettongia", "penicillata", "442b9048417ef20cf680a0ae26ee4d86")

}
\references{
Scott Chamberlain, Eduard Szocs (2013). "taxize - taxonomic search and retrieval in R." F1000Research. http://f1000research.com/articles/2-191/v2.
}
