% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{FetchSpT}
\alias{FetchSpT}
\title{FetchSpT wrapper}
\usage{
FetchSpT(db, genus, species, synonyms, additionalkeywords, language = 0)
}
\arguments{
\item{db}{Literature database. Scopus ("scopus") or Web of Science ("wos").}

\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{synonyms}{Alternate species names.}

\item{additionalkeywords}{Optional search terms.}

\item{language}{Language of the paper; default is 0, enter 1 to retrieve the variable. Scopus only.}
}
\value{
A dataframe of the species' citation records with the given \code{genus} and \code{species}.
}
\description{
This is a wrapper function for \code{\link{FetchSpT_scopus}}, and \code{\link{FetchSpT_wos}}.
}
\examples{
\dontrun{
FetchSpT("scopus", genus = "Osphranter", species = "rufus")
}
\dontrun{
FetchSpT("scopus", genus = "Osphranter", species = "rufus", synonyms = "Macropus rufus", additionalkeywords = "conserv*")
}
}
