% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{FetchSpTAK}
\alias{FetchSpTAK}
\title{Fetch data - title, abstract and keywords}
\usage{
FetchSpTAK(genus, species, APIkey)
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{APIkey}{Scopus API key needed to access and download data from their database.}
}
\value{
A dataframe of the species' citation records with the given \code{genus} and \code{species}.
}
\description{
This function fetches citation information from Scopus using genus and species name found in the title, abstract and keywords of the publications.
Duplicates are removed after fetching the data.
}
\examples{
\dontrun{
FetchSpT("Bettongia", "penicillata", "442b9048417ef20cf680a0ae26ee4d86")

#lower case letter in genus is also accepted and will return identical results

FetchSpT("bettongia", "penicillata", "442b9048417ef20cf680a0ae26ee4d86")
}
}
