% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{FetchSpTAK}
\alias{FetchSpTAK}
\title{Fetch data - title, abstract and keywords}
\usage{
FetchSpTAK(genus, species, synonyms, additionalkeywords, APIkey)
}
\arguments{
\item{genus}{Genus classification from the binomial name.}

\item{species}{Species classification from the binomial name.}

\item{synonyms}{Alternate species names}

\item{additionalkeywords}{Optional search terms.}

\item{APIkey}{Scopus API key needed to access and download data from their database.}
}
\value{
A dataframe of the species' citation records with the given \code{genus} and \code{species}.
}
\description{
This function fetches citation information from Scopus using genus and species name found in the title, abstract and keywords of the publications.
Duplicates are removed after fetching the data.
}
\examples{
\dontrun{
FetchSpTAK("Bettongia", "penicillata", APIkey = "myAPI")

#lower case letter in genus is also accepted and will return identical results

FetchSpTAK("bettongia", "penicillata", APIkey = "myAPI")
}
\dontrun{
FetchSpTAK("Bettongia", "penicillata", "conserv*", "myAPI")

#lower case letter in genus is also accepted and will return identical results

FetchSpTAK("bettongia", "penicillata", "conserv*", "myAPI")
}
}
