% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koala.R
\docType{data}
\name{Koala}
\alias{Koala}
\title{Koala dataset}
\format{
A data frame with 773 rows and 20 variables
}
\source{
\url{http://api.elsevier.com/content/search/scopus}
}
\usage{
Koala
}
\description{
Citation records of koala (Phascolarctos cinereus) from Scopus.
Data was retrieved on 10 July 2020.
}
\keyword{datasets}
