% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\docType{data}
\name{languages}
\alias{languages}
\title{Languages}
\format{
A csv file with the complete list of languages of documents found on Scopus.
}
\source{
\url{https://www.elsevier.com/solutions/scopus/how-scopus-works/content}
}
\usage{
languages
}
\description{
List of languages of documents found on Scopus.
Data was retrieved on 9 March 2021.
}
\keyword{datasets}
