% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{plotAllindices}
\alias{plotAllindices}
\title{Index plot}
\usage{
plotAllindices(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{Allindices}}.}
}
\value{
ggplot
}
\description{
Plots the indices of a single species or combined.
}
\examples{
\dontrun{
Woylie <- Fetch(db = "scopus",
                search = "tak",
                genus = "Bettongia",
                species = "penicillata")
Quokka <- Fetch(db = "scopus",
                search = "tak",
                genus = "Setonix",
                species = "brachyurus")
Platypus <- Fetch(db = "scopus",
                  search = "tak",
                  genus = "Ornithorhynchus",
                  species = "anatinus")
Koala <- Fetch(db = "scopus",
               search = "tak",
               genus = "Phascolarctos",
               species = "cinereus")
}
W <- Allindices(Woylie,
                genus = "Bettongia",
                species = "penicillata")
Q <- Allindices(Quokka,
                genus = "Setonix",
                species = "brachyurus")
P <- Allindices(Platypus,
                genus = "Ornithorhynchus",
                species = "anatinus")
K <- Allindices(Koala,
                genus = "Phascolarctos",
                species = "cinereus")
CombineSp <- dplyr::bind_rows(W, Q, P, K)
plotAllindices(CombineSp)

}
