% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{SpH5}
\alias{SpH5}
\title{Species h5 index}
\usage{
SpH5(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
H5 index.
}
\description{
This function calculates the h-index of a species in the past 5 years.
}
\examples{
data(Woylie)
SpH5(Woylie)

}
\references{
Suzuki, H. (2012). Google Scholar Metrics for Publications. Retrieved from https://scholar.googleblog.com/2012/04/google-scholar-metrics-for-publications.html.
}
