% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{SpHindex}
\alias{SpHindex}
\title{Species h-index}
\usage{
SpHindex(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
H-index.
}
\description{
This function calculates the h-index of a species.
}
\examples{
data(Woylie)
SpHindex(Woylie)

}
\references{
Bertoli-Barsotti, L. & Lando, T. (2015). On a formula for the h-index. \emph{Journal of Informetrics, 9}(4), 762-776. \cr
Hirsch, J. (2005). An index to quantify an individual's scientific research output. \emph{Proceedings of the National Academy of Sciences of the United States of America, 102}(46), 16569-16572.
}
