% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{Spi10}
\alias{Spi10}
\title{Species i10 index}
\usage{
Spi10(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
i10 index.
}
\description{
This function calculates the i10 index of a species.
i10 index counts all of the publications with 10 or more citations.
}
\examples{
data(Woylie)
Spi10(Woylie)

}
\references{
Cornell University (2019). i10-index. Retrieved from https://guides.library.cornell.edu/c.php?g=32272&p=203393.
}
