% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{SpMindex}
\alias{SpMindex}
\title{Species m-index}
\usage{
SpMindex(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{Fetch}}.}
}
\value{
M-index.
}
\description{
This function calculates the m-index of  species.
M-index uses the h-index and divides it by the number of years of activity.
}
\examples{
\dontrun{
Woylie <- Fetch(db = "scopus",
                search = "tak",
                genus = "Bettongia",
                species = "penicillata")
}
SpMindex(Woylie)

}
\references{
University of Pittsburgh (2019). \emph{Research Impact and Metrics: Author metrics.} Retrieved from \url{https://pitt.libguides.com/bibliometricIndicators/AuthorMetrics}.
}
