% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{SpMindex}
\alias{SpMindex}
\title{Species m-index}
\usage{
SpMindex(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
M-index.
}
\description{
This function calculates the m-index of  species.
M-index uses the h-index and divides it by the number of years of activity.
}
\examples{
SpMindex(data)

}
