% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{TotalCite}
\alias{TotalCite}
\title{Total citations}
\usage{
TotalCite(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
A numerical value of the total number of citations.
}
\description{
This function calculates the total number of citations.
}
\examples{
data(Woylie)
TotalCite(Woylie)

}
