% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{TotalJournals}
\alias{TotalJournals}
\title{Total journals}
\usage{
TotalJournals(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{Fetch}}.}
}
\value{
An integer of the total number of journals.
}
\description{
This function calculates the total number of journals.
}
\examples{
\dontrun{
Woylie <- Fetch(db = "scopus",
                search = "tak",
                level = "species",
                genus = "Bettongia", species = "penicillata")
}
TotalJournals(Woylie)

}
