% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieshindex.R
\name{TotalPub}
\alias{TotalPub}
\title{Total publications}
\usage{
TotalPub(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
An integer of the total number of publications.
}
\description{
This function calculates the total number of publications.
}
\examples{
data(Woylie)
TotalPub(Woylie)

}
