% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hindex.R
\name{TotalRev}
\alias{TotalRev}
\title{This function calculates the total number of reviews.}
\usage{
TotalRev(data)
}
\arguments{
\item{data}{The dataframe generated from \code{\link{FetchSpT}} or \code{\link{FetchSpTAK}}.}
}
\value{
An integer of the total number of reviews.
}
\description{
This function calculates the total number of reviews.
}
\examples{
TotalRev(SpeciesData)

}
